# PERF: use new to_records() argument in to_stata()