@bashtage 2019-01-31
Code changes seem fine. Might need a benchmark to show the gains on strings though since these seem to be time conversion tests.